reports work well and code moved to different files
