+++ app/app.py
@@ -54,25 +54,26 @@ def get_example_use(software_name):
         print(e)
         return(jsonify({"use": '**Unable to find use case record**'})), 500
 
-@app.route('/report-issue', methods=['POST'])
+@app.route("/report-issue", methods=['POST'])
 def report_issue():
     issue_report = request.get_json()
 
-    report, captureDataUrl = sanitize_and_process_reports(issue_report)
-
+    report = sanitize_and_process_reports(issue_report)
     current_datetime = report['datetime']
 
-    report_folder = os.path.join('reports',current_datetime)
-    os.makedirs(report_folder, exist_ok=True)
+    capture_data_url = report['captureDataUrl']
+    report.pop('captureDataUrl')
 
-    report_filename = os.path.join(report_folder,'report.json')
+    report_folder = os.path.join('reports', current_datetime)
+    os.makedirs(report_folder, exist_ok=True)
+    report_filename = os.path.join(report_folder, 'report.json')
     with open(report_filename, 'w') as f:
         json.dump(report, f, indent=4)
 
-    capture_data_url = urlopen(captureDataUrl)
+    capture_data = urlopen(capture_data_url).read()
     capture_filename = os.path.join(report_folder, report['captureFilename'])
     with open(capture_filename, 'wb') as f:
-        f.write(capture_data_url.file.read())
+        f.write(capture_data)
 
     return jsonify({'message': 'Issue reported successfully'})
