+++ app/templates/base.html
@@ -57,12 +57,26 @@
                         <div class="col">
                             <ul class="nav nav-pills">
                                 <li class="nav-item">
-                                  <a class="nav-link" aria-current="page" href="/">SDS Static</a>
+                                  <a class="nav-link" id="static-link" aria-current="page" href="/">SDS Static</a>
                                 </li>
                                 <li class="nav-item">
-                                  <a class="nav-link" href="/dynamic">SDS Dynamic</a>
+                                  <a class="nav-link" id="dynamic-link" href="/dynamic">SDS Dynamic</a>
                                 </li>
                             </ul>
+                        </div>
+                        <div class="col text-right">
+                            <div class="btn-group">
+                                <button id="reportIssueBtn" class="btn btn-outline-danger">
+                                    <span id="reportIssueText">Report Issue</span> 
+                                    <img width="48" height="48" src="https://img.icons8.com/fluency-systems-regular/48/webpage-click.png" alt="webpage-click"/>
+                                </button>
+                                <button id="customReportBtn" class="btn btn-outline-danger">
+                                    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pencil-square" viewBox="0 0 16 16">
+                                        <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z"/>
+                                        <path fill-rule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5z"/>
+                                    </svg>
+                                </button>
+                            </div>
                         </div>
                     </div>
                     <div class="row">
