+++ app/templates/reportModal.html
@@ -6,7 +6,7 @@
           <button type="button" class="close" data-dismiss="modal">&times;</button>
         </div>
         <div id="reportBody" class="modal-body">
-          <p>Please provide any additional details about the issue:</p>
+          <p>Please describe the issue:</p>
           <form>
             <div class="form-group">
                 <label for="reportFeedback">Feedback:</label>
