+++ app/templates/software_search.html
@@ -8,7 +8,9 @@
     <script src="https://cdn.jsdelivr.net/npm/showdown@1.9.1/dist/showdown.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.4.0/highlight.min.js"></script>
     <script src="https://html2canvas.hertzen.com/dist/html2canvas.min.js"></script>
-    <script type="module" src="../static/js/table.js"></script>
+    <script type="module" src="{{url_for('static', filename='js/alerts.js')}}"></script>
+    <script type="module" src="{{url_for('static', filename='js/table.js')}}"></script>
+    <script type="module" src="{{url_for('static', filename='js/issueReporting.js')}}"></script>
 
 {% endblock %}
 <div>
@@ -39,8 +41,6 @@
     <label><strong>Hover your mouse to the edge of the table to scroll.</strong></label>
 </div>
 {%endif%}
-
-<button id="reportIssueBtn" class="btn btn-danger">Report Issue</button>
 
 <div class="table-responsive table-container">
     <div class="table-responsive" id="softwareTableDiv" style="width: 80vw;">
